+++ internal/db/matches.go
@@ -1,7 +1,8 @@
 package db
 
 import (
-	"github.com/google/uuid"
+	"uuid"
+
 	"gorm.io/gorm"
 )
 
@@ -15,7 +16,7 @@ type Match struct {
 
 type MatchParticipant struct {
 	MatchID   uint      `gorm:"primaryKey;autoIncrement:false"`
-	UserID    uuid.UUID `gorm:"primaryKey;autoIncrement:false"`
+	UserID    uuid.UUID `gorm:"primaryKey;autoIncrement:false;serializer:stduuid"`
 	Placement int       // 1 for first place/winner, 2 for second, etc.
 	EloDelta  int       // How much elo they gained/lost